Clase 9 - Video 9.1
Instalación de Intellij y abrimos el primer archivo.
