+++ Java/Semana 4 (24-04 al 28-04)/Leccion03/ForEach/src/domain/Persona.java
