+++ Java/Semana 4 (24-04 al 28-04)/Leccion03/ModificadoresAcceso/src/paquete1/Clase2.java
